add booking apis and fix relationship, attributes

+++ SkinCareBookingSystem/DTOs/ServiceDTO.cs
@@ -6,7 +6,7 @@
         public int ServiceCategoryId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public TimeSpan Duration { get; set; }
         public string VideoURL { get; set; } = string.Empty;
         public float Rating { get; set; }
@@ -19,7 +19,7 @@
         public int ServiceCategoryId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public TimeSpan Duration { get; set; }
         public string VideoURL { get; set; } = string.Empty;
         public float Rating { get; set; }
@@ -32,7 +32,7 @@
         public int ServiceCategoryId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public TimeSpan Duration { get; set; }
         public string VideoURL { get; set; } = string.Empty;
         public float Rating { get; set; }
